+++ server/src/save.py
@@ -13,7 +13,7 @@ When writing migrations make sure that these things are respected:
     - e.g. a column added to Circle also needs to be added to CircularToken
 """
 
-SAVE_VERSION = 72
+SAVE_VERSION = 73
 
 import json
 import logging
@@ -177,6 +177,19 @@ def upgrade(db: SqliteExtDatabase, version: int):
         with db.atomic():
             db.execute_sql(
                 "ALTER TABLE user ADD COLUMN colour_history TEXT DEFAULT NULL"
+            )
+    elif version == 72:
+        # Change default zoom level from 1.0 to 0.2
+        with db.atomic():
+            db.execute_sql(
+                "CREATE TEMPORARY TABLE _location_user_option_72 AS SELECT * FROM location_user_option"
+            )
+            db.execute_sql("DROP TABLE location_user_option")
+            db.execute_sql(
+                'CREATE TABLE IF NOT EXISTS "location_user_option" ("id" INTEGER NOT NULL PRIMARY KEY, "location_id" INTEGER NOT NULL, "user_id" INTEGER NOT NULL, "pan_x" REAL DEFAULT 0 NOT NULL, "pan_y" REAL DEFAULT 0 NOT NULL, "zoom_display" REAL DEFAULT 0.2 NOT NULL, "active_layer_id" INTEGER, FOREIGN KEY ("location_id") REFERENCES "location" ("id") ON DELETE CASCADE, FOREIGN KEY ("user_id") REFERENCES "user" ("id") ON DELETE CASCADE, FOREIGN KEY ("active_layer_id") REFERENCES "layer" ("id"));'
+            )
+            db.execute_sql(
+                'INSERT INTO "location_user_option" (id, location_id, user_id, pan_x, pan_y, zoom_display, active_layer_id) SELECT id, location_id, user_id, pan_x, pan_y, zoom_display, active_layer_id FROM _location_user_option_72'
             )
     else:
         raise UnknownVersionException(