+++ CHANGELOG.md
@@ -26,6 +26,7 @@ tech changes will usually be stripped from release notes for the public
 -   Defeat cross now scales better with shape size
 -   Shape badge now scales better with shape size
 -   Show default "no" button in delete/leave campaign prompt
+-   Default location zoom level is now 0.2 instead of 1.0
 -   [server] Added log rotation
 -   [server] Restructure server files
 -   [tech] SyncTo primitive modified to an alternative Sync structure